+++ app/controller/StudentController.php
@@ -1,0 +1,13 @@
+<?php
+class StudentController {
+    private $studentModel = NULL;
+
+    public function __construct() {
+        $this->studentModel = new StudentModel();
+    }
+
+    public function handleRequest() {
+        echo $this->studentModel->getRequirementsForStudent();
+    }
+}
+?>
