+++ app/model/PlanModel.php
@@ -1,88 +1,90 @@
 <?php
 
-require_once(__DIR__."/../config/config.php");
-
-class PlanModel {
+class PlanModel
+{
     private $conn = NULL;
 
-    public function __construct() {
+    public function __construct()
+    {
         $this->conn = new PDO(DBCONNECTSTRING, DBUSER, DBPASSWORD);
     }
 
-    //UPDATED DONE USED
-    public function movePlanItem($token, $studentId, $courseId, $semester,
-                                 $year, $toSemester, $toYear, $reqId = null)
+    public function createPlan()
     {
-        try {
-           // if (!validateToken($token, $studentId)) {
-           //     return 403;
-           // }
+        $errormsg = "";
+        $showform = 1;
 
-            //  if(empty($studentId)) return 404;
+            // Cleanse title on submission
+            $formfield['title'] = htmlspecialcharacters(stripslashes(trim($_POST['title'])));
 
-            $sql = 'UPDATE course_records SET semesterCode=:toSemester, year=:toYear WHERE studentId=:studentId AND courseId= :courseId AND semesterCode=:semester AND year=:year AND type=2';
-            //$sql = $sql. ' VALUES (null, :studentId, :courseId, null, :semester, :year, :reqId, 2, :proposedReqId)';
-            $stmt = $this->conn->prepare($sql);
+            // Check if title name is empty
+            if (empty($formfield['title'])) {
+                $errormsg .= "<p>Title is empty</p>";
+            }
 
-            $stmt->bindParam(':studentId', $studentId);
-            $stmt->bindParam(':semester', $semester);
-            $stmt->bindParam(':year', $year);
-            $stmt->bindParam(':courseId', $courseId);
-            $stmt->bindParam(':toSemester', $toSemester);
-            $stmt->bindParam(':toYear', $toYear);
+            // Check for duplicate title
+            if ($formfield['title'] != $_POST['origtitle']) {
+                try {
+                    // Pulls titles from the database & binds value to variable to be used
+                    $sqltitle = 'SELECT * FROM '.DBNAME.' WHERE title = :title';
+                    $stmttitle = $this->conn->prepare($sqltitle);
+                    $stmttitle->bindValue(':title', $formfield['title']);
+                    $stmttitle->execute();
+                    $count = $stmttitle->rowCount();
 
-            $success = $stmt->execute();
-            $inserted = $success ? "yes" : "no";
-            echo "<h4>success:".$inserted."</h4>";
+                   //fix counttitle not declared???
+                   // if ($counttitle > 0) {
+                   //     $errormsg .= "<p>Duplicate plan name.</p>";
+                   // }
+                } catch (PDOException $e) {
+                    echo 'Unable to fetch title to check for existing. <br />ERROR: <br />' . $e->getMessage();
+                    exit();
+                }
+            }
 
+            // Update if no errors exist
+            if ($errormsg != "") {
+                echo $errormsg;
+                echo "<p>Try again.</p>";
+            } else {
+                try {
+                    // Insert data into database
+                    $sqlupdate = 'UPDATE '.DBNAME.' SET title = :title WHERE ID = :ID';
+                    $stmtupdate = $this->conn->prepare($sqlupdate);
+                    $stmtupdate->bindValue(':title', $formfield['title']);
+                    $stmtupdate->bindValue(':ID', $_POST['x']);
+                    $stmtupdate->execute();
 
-        }//end try
-        catch (PDOException $e) {
-            return $sql . "<br>" . $e->getMessage();
-            //return 500;
+                    // Hide form
+                    $showform = 0;
+                } catch (PDOException $e) {
+                    echo 'Error updating title <br />ERROR: <br />' . $e->getMessage();
+                    exit();
+                }
+            }
+
+            if ($showform == 1) {
+                try {
+                    // Pull data from database for existing plans
+                    $sql = 'SELECT * FROM '.DBNAME.' WHERE ID = :ID';
+
+                    $stmt = $this->conn->prepare($sql);
+                    $stmt->bindValue(':ID', $_GET['x']);
+                    $stmt->execute();
+                    $row = $stmt->fetch();
+                    ?>
+                    <!-- Form for changing plan title -->
+                    <form action="#" id="titleForm" method="post" name="titleForm">
+                        <input id="titleName" name="name" placeholder="Name" type="text">
+                        <input type="submit" id="changeTitle" onclick="titleSubmit();" value="Submit">
+                    </form>
+                    <!-- End form -->
+                    <?php
+                } catch (PDOException $e) {
+                    echo 'Error fetching plans. <br />ERROR: </br>' . $e->getMessage();
+                    exit();
+                }
+            }
         }
-
-        $conn = null;
-        return $inserted;
     }
-
-    function addPlanItem($token, $studentId, $courseId, $semester, $year, $reqId = null, $proposedReqId = null)
-    {
-        echo "in addPlanItem";
-        try {
-            //  if(!validateToken($token, $studentId)) {return 403;}
-
-            //  if(empty($studentId)) return 404;
-
-            $conn = new PDO(DBCONNECTSTRING, DBUSER, DBPASSWORD);
-            $sql = 'INSERT INTO course_records (id, studentId, courseId, grade, semesterCode, year, reqId, type, proposedReqId) ';
-            $sql = $sql . ' VALUES (null, :studentId, :courseId, null, :semester, :year, :reqId, 2, :proposedReqId)';
-            $stmt = $conn->prepare($sql);
-
-            echo "courseid: " . $courseId;
-
-            $stmt->bindParam(':studentId', $studentId);
-            $stmt->bindParam(':semester', $semester);
-            $stmt->bindParam(':year', $year);
-            $stmt->bindParam(':courseId', $courseId);
-            $stmt->bindParam(':reqId', $reqId);
-            $stmt->bindParam(':proposedReqId', $proposedReqId);
-            $success = $stmt->execute();
-            $inserted = $success ? "yes" : "no";
-            $result = "<h4>success:" . $inserted . "</h4>";
-            echo $result;
-
-
-        }//end try
-        catch (PDOException $e) {
-            echo $sql . "<br>" . $e->getMessage();
-            return 500;
-        }
-
-        $conn = null;
-        return $result;
-
-    }
-
-}
 ?>
